have added a new controller to hold the revneuse APIs
