+++ balanced-car-api/app/controllers/api/V1/users_controller.rb
@@ -28,19 +28,6 @@ class Api::V1::UsersController < ApplicationController
     render json: @bills
   end
 
-  def index_revenues
-    @revenues = User.includes(:revenues).find_by_id(current_devise_api_token.resource_owner)
-    render json: @revenues.revenues
-  end
-
-  def index_car_revenues
-    user_id = current_devise_api_token.resource_owner.id
-    car_id = params[:id]
-
-    @revenues = Revenue.joins(:car).where(cars: { id: car_id, user_id: })
-    render json: @revenues
-  end
-
   def show_car
     @car = Car.find_by_id(params[:id])
     check_user(@car.user)
@@ -57,12 +44,6 @@ class Api::V1::UsersController < ApplicationController
     @bill = Bill.find_by_id(params[:id])
     check_user(@bill.user)
     render json: @bill
-  end
-
-  def show_revenue
-    @revenue = Revenue.find_by_id(params[:id])
-    check_user(@revenue.user)
-    render json: @revenue
   end
 
   def switch_publicity
