+++ balanced-car-api/test/controllers/api/v1/revenues_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class Api::V1::RevenuesControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
