+++ balanced-car-api/test/controllers/revenues_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class RevenuesControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
